ClearLogo
it will now use art directly instead of using cache

+++ include.php
@@ -184,6 +184,8 @@ function displaybanner ($filename, $bannerauthor, $allowdelete, $fullurl, $banne
 
 ## Function to generate/display a cached version of a banner
 function displaybannernew ($banner, $allowdelete, $link) {
+    if(substr($banner->filename,0,9) == 'clearlogo') $clogo = 1;
+    else $clogo = 0;
 
 	global $loggedin, $user, $bannercount, $tab, $id, $seriesid, $seasonid, $baseurl;
 	switch ($tab) {
@@ -199,7 +201,7 @@ function displaybannernew ($banner, $allowdelete, $link) {
 
 
 	## Check if the banner is cached already. If not, create it.
-	if (!file_exists("../banners/_cache/$banner->filename"))  {
+	if (!file_exists("../banners/_cache/$banner->filename") && !$clogo)  {
 		## Create the cached version of the image
 		$target = file_get_contents("$baseurl/thumbnail.php?gd=2&maxw=300&src=banners/$banner->filename");
 		file_put_contents("banners/_cache/$banner->filename", $target);
@@ -210,12 +212,15 @@ function displaybannernew ($banner, $allowdelete, $link) {
 		$imgHeight = $image_height;
 	}
 
-
-	## Display the image
-	if ($link)  {
+    ## Display the image
+	
+	if($clogo && $link) {
+		print "<a href=\"$link\"><img src=\"$baseurl/banners/$banner->filename\" class=\"banner\" width=\"$imgWidth\" height=\"$imgHeight\" border=\"0\" alt=\"$banner->seriesname\"></a>\n";
+	}
+    if ($link)  {
 		print "<a href=\"$link\"><img src=\"$baseurl/banners/_cache/$banner->filename\" class=\"banner\" width=\"$imgWidth\" height=\"$imgHeight\" border=\"0\" alt=\"$banner->seriesname\"></a>\n";
 	}
-	else  {
+	else {
 		print "<img src=\"$baseurl/banners/_cache/$banner->filename\" class=\"banner\" width=\"$imgWidth\" height=\"$imgHeight\" border=\"0\">\n";
 	}
